modified:   src/main/java/TP/Barrier.java 	modified:   src/main/java/TP/Buffer.java 	modified:   src/main/java/TP/Main.java 	modified:   src/main/java/TP/PerfectWorker.java 	modified:   src/main/java/TP/ReportFactory.java 	modified:   src/main/java/TP/ThreadPool.java 	modified:   target/classes/META-INF/MANIFEST.MF 	modified:   target/classes/META-INF/maven/TrabajoPracticoAutoReporte/TrabajoPracticoAutoReporte/pom.properties 	modified:   target/classes/TP/Barrier.class 	modified:   target/classes/TP/Buffer.class 	modified:   target/classes/TP/Main.class 	modified:   target/classes/TP/PerfectWorker.class 	modified:   target/classes/TP/ThreadPool.class 	src/main/java/TP/MonitorResultado.java 	target/classes/TP/MonitorResultado.class

+++ src/main/java/TP/Buffer.java
@@ -1,8 +1,10 @@
 package TP;
 
+import java.math.BigInteger;
+
 public class Buffer extends Thread{
 	
-	long[] buffer;
+	BigInteger[] buffer;
 	int size;
 	//Para el pull
 	int currentPosition = 0;
@@ -11,11 +13,11 @@ public class Buffer extends Thread{
 	int nextAdd = 0;
 	
 	public Buffer(int n) {
-		this.buffer = new long[n];
+		this.buffer = new BigInteger[n];
 		this.size = n;
 	}
 
-	public synchronized void push(long n) {
+	public synchronized void push(BigInteger n) {
 		while(bufferLleno()) {
 			try {
 				wait();
@@ -29,7 +31,7 @@ public class Buffer extends Thread{
 		notifyAll();
 	}
 
-	public synchronized long get(){
+	public synchronized BigInteger get(){
 		while(bufferVacio()) {
 			try {
 				wait();
@@ -37,7 +39,7 @@ public class Buffer extends Thread{
 				e.printStackTrace();
 			}
 		}
-		long a = buffer[currentPosition];
+		BigInteger a = buffer[currentPosition];
 		currentPosition++;
 		notifyAll();
 		return a;
